# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APTLY",
    "description": "Evidence-Grounded Multimodal AI Interview Coach \u2014 REST API\n\nPhase 0: Foundation scaffold. See /docs for OpenAPI specification.\nSee docs/api/contracts.md for planned endpoint contracts.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/interviews": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Create and configure interview",
        "description": "Creates an interview session and dynamically generates tailored questions.",
        "operationId": "create_interview_api_v1_interviews_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "Get interview details",
        "description": "Fetches an interview, its active state, generated questions, and answers.",
        "operationId": "get_interview_api_v1_interviews__interview_id__get",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/start": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Start interview session",
        "description": "Transitions the interview to active/running state and activates Question 1.",
        "operationId": "start_interview_api_v1_interviews__interview_id__start_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/answers": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Create answer record",
        "description": "Initializes a candidate answer recording for a question.",
        "operationId": "create_answer_api_v1_interviews__interview_id__answers_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/answers/{answer_id}/upload": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Upload answer audio",
        "description": "Uploads binary audio recording (WebM/WAV) and triggers async speech processing.",
        "operationId": "upload_answer_audio_api_v1_interviews__interview_id__answers__answer_id__upload_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          },
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_answer_audio_api_v1_interviews__interview_id__answers__answer_id__upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/next-question": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Advance to next question",
        "description": "Advances the interview to the next question index or marks it completed.",
        "operationId": "next_question_api_v1_interviews__interview_id__next_question_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/finish": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Finish interview session",
        "description": "Finalizes the interview and marks it complete.",
        "operationId": "finish_interview_api_v1_interviews__interview_id__finish_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}/review": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "Get post-interview review",
        "description": "Returns comprehensive speech metrics, transcripts, and question-by-question breakdown.",
        "operationId": "get_interview_review_api_v1_interviews__interview_id__review_get",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerCreateRequest": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "AnswerCreateRequest",
        "description": "Payload to initialize a candidate answer for a question."
      },
      "AnswerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "interview_id": {
            "type": "string",
            "format": "uuid",
            "title": "Interview Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "sequence_number": {
            "type": "integer",
            "title": "Sequence Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "audio_storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Storage Key"
          },
          "audio_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Size Bytes"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "speech_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeechMetricsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "interview_id",
          "question_id",
          "sequence_number",
          "status",
          "duration_seconds",
          "created_at"
        ],
        "title": "AnswerResponse",
        "description": "Candidate answer entity."
      },
      "Body_upload_answer_audio_api_v1_interviews__interview_id__answers__answer_id__upload_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File",
            "description": "Binary audio recording file"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds",
            "description": "Total recorded duration in seconds",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_upload_answer_audio_api_v1_interviews__interview_id__answers__answer_id__upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewCreateRequest": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id",
            "description": "Associated job ID if created from a JD."
          },
          "role_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Profile Id",
            "description": "Associated role profile ID."
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title",
            "default": "Practice Interview"
          },
          "interview_type": {
            "type": "string",
            "title": "Interview Type",
            "description": "mixed, technical, behavioral",
            "default": "mixed"
          },
          "difficulty_level": {
            "type": "string",
            "title": "Difficulty Level",
            "description": "easy, medium, hard",
            "default": "medium"
          },
          "target_duration_minutes": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 3.0,
            "title": "Target Duration Minutes",
            "default": 10
          },
          "question_count": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Question Count",
            "default": 3
          }
        },
        "type": "object",
        "title": "InterviewCreateRequest",
        "description": "Payload to create and configure a new interview session."
      },
      "InterviewDetailResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "description": "Schema version. Increment when structure changes.",
            "default": "1.0",
            "examples": [
              "1.0",
              "1.1",
              "2.0"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "interview_type": {
            "type": "string",
            "title": "Interview Type"
          },
          "difficulty_level": {
            "type": "string",
            "title": "Difficulty Level"
          },
          "target_duration_minutes": {
            "type": "integer",
            "title": "Target Duration Minutes"
          },
          "current_question_index": {
            "type": "integer",
            "title": "Current Question Index"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "role_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerResponse"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "interview_type",
          "difficulty_level",
          "target_duration_minutes",
          "current_question_index",
          "created_at"
        ],
        "title": "InterviewDetailResponse",
        "description": "Full detail view of an Interview including questions and answers."
      },
...
      "InterviewReviewResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "description": "Schema version. Increment when structure changes.",
            "default": "1.0",
            "examples": [
              "1.0",
              "1.1",
              "2.0"
            ]
          },
          "interview": {
            "$ref": "#/components/schemas/InterviewResponse"
          },
          "role_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_duration_seconds": {
            "type": "number",
            "title": "Total Duration Seconds"
          },
          "total_answers_count": {
            "type": "integer",
            "title": "Total Answers Count"
          },
          "average_wpm": {
            "type": "number",
            "title": "Average Wpm"
          },
          "total_fillers_count": {
            "type": "integer",
            "title": "Total Fillers Count"
          },
          "overall_filler_density": {
            "type": "number",
            "title": "Overall Filler Density"
          },
          "total_pauses_count": {
            "type": "integer",
            "title": "Total Pauses Count"
          },
          "questions_review": {
            "items": {
              "$ref": "#/components/schemas/QuestionReviewItem"
            },
            "type": "array",
            "title": "Questions Review"
          }
        },
        "type": "object",
        "required": [
          "interview",
          "total_duration_seconds",
          "total_answers_count",
          "average_wpm",
          "total_fillers_count",
          "overall_filler_density",
          "total_pauses_count"
        ],
        "title": "InterviewReviewResponse",
        "description": "Comprehensive post-interview review view."
      },
...
      "RoleProfileResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "description": "Schema version. Increment when structure changes.",
            "default": "1.0",
            "examples": [
              "1.0",
              "1.1",
              "2.0"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "role_title": {
            "type": "string",
            "title": "Role Title"
          },
          "seniority": {
            "type": "string",
            "title": "Seniority"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "technical_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technical Skills"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          },
          "responsibilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Responsibilities"
          },
          "behavioral_competencies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Behavioral Competencies"
          },
          "interview_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interview Topics"
          },
          "preferred_experience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Experience"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version",
            "default": "v1"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "role_title",
          "seniority",
          "domain",
          "created_at"
        ],
        "title": "RoleProfileResponse",
        "description": "Structured role profile extracted from a job description."
      },
...
    }
  }
}